[FIX] Remove typecast of char_t in calls to std::tolower/toupper
Also changed "default" locale to "en_US.utf8".

+++ eonregex/sources/CharGroup.cpp
@@ -29,8 +29,8 @@ namespace eon
 				if( data.flags() & Flag::icase )
 				{
 					auto& chars = Characters::get();
-					auto l = std::tolower( static_cast<int>( data() ), chars.locale() );
-					auto u = std::toupper( static_cast<int>( data() ), chars.locale() );
+					auto l = std::tolower( data(), chars.locale() );
+					auto u = std::toupper( data(), chars.locale() );
 					if( l != u )
 					{
 						if( data() == l )
